refactor(terser): no longer a nested workspace
doesn't get installed into a funny @npm_bazel_terser package for users either
