+++ examples/app/BUILD.bazel
@@ -1,9 +1,9 @@
 load("@build_bazel_rules_nodejs//:index.bzl", "pkg_web")
+load("@npm//@bazel/terser:index.bzl", "terser_minified")
 load("@npm//html-insert-assets:index.bzl", "html_insert_assets")
 load("@npm//http-server:index.bzl", "http_server")
 load("@npm_bazel_protractor//:index.bzl", "protractor_web_test_suite")
 load("@npm_bazel_rollup//:index.bzl", "rollup_bundle")
-load("@npm_bazel_terser//:index.bzl", "terser_minified")
 load("@npm_bazel_typescript//:index.bzl", "ts_devserver", "ts_library", "ts_project")
 
 package(default_visibility = ["//visibility:public"])
