+++ internal/npm_install/test/npm_packages/spec.js
@@ -1,11 +1,10 @@
 const fs = require('fs');
-const path = require('path');
+const runfiles = require(process.env['BAZEL_NODE_RUNFILES_HELPER']);
 
 describe('installing hybrid packages', () => {
   it('should work', () => {
-    const content = fs.readFileSync(
-        path.join(process.env['TEST_SRCDIR'], 'npm', 'bazel_workspaces_consistent', 'a.txt'),
-        'utf-8');
+    const content =
+        fs.readFileSync(runfiles.resolve('npm/bazel_workspaces_consistent/a.txt'), 'utf-8');
     expect(content).toEqual('some content');
   });
 });
