+++ internal/pkg_web/test/BUILD.bazel
@@ -2,8 +2,8 @@ load("@build_bazel_rules_nodejs//:index.bzl", "pkg_web")
 load("@npm//@babel/cli:index.bzl", "babel")
 load("@npm_bazel_jasmine//:index.bzl", "jasmine_node_test")
 load("@npm_bazel_rollup//:index.bzl", "rollup_bundle")
-load("@npm_bazel_terser//:index.bzl", "terser_minified")
 load("@npm_bazel_typescript//:index.bzl", "ts_library")
+load("//packages/terser:index.bzl", "terser_minified")
 
 package(default_visibility = ["//visibility:public"])
