+++ packages/protractor/test/protractor-2/BUILD.bazel
@@ -1,8 +1,8 @@
 load("@npm//http-server:index.bzl", "http_server")
 load("@npm_bazel_protractor//:index.bzl", "protractor_web_test_suite")
 load("@npm_bazel_rollup//:index.bzl", "rollup_bundle")
-load("@npm_bazel_terser//:index.bzl", "terser_minified")
 load("@npm_bazel_typescript//:index.bzl", "ts_config", "ts_devserver", "ts_library")
+load("//packages/terser:index.bzl", "terser_minified")
 
 ts_config(
     name = "tsconfig-test",
