+++ packages/terser/BUILD.bazel
@@ -12,8 +12,32 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 load("@build_bazel_rules_nodejs//:tools/defaults.bzl", "pkg_npm")
+load("@build_bazel_rules_nodejs//tools/stardoc:index.bzl", "stardoc")
 load("//third_party/github.com/bazelbuild/bazel-skylib:rules/copy_file.bzl", "copy_file")
+
+package(default_visibility = ["//visibility:public"])
+
+exports_files(["terser_config.default.json"])
+
+bzl_library(
+    name = "bzl",
+    srcs = glob(["*.bzl"]),
+    deps = [
+        "@build_bazel_rules_nodejs//:bzl",
+        "@build_bazel_rules_nodejs//internal/common:bzl",
+    ],
+)
+
+stardoc(
+    name = "docs",
+    testonly = True,
+    out = "index.md",
+    input = "index.bzl",
+    tags = ["fix-windows"],
+    deps = [":bzl"],
+)
 
 # Ugly genrule depending on local linux environment to build the README out of skylark doc generation.
 # Only referenced when we do a release.
@@ -21,11 +45,11 @@ load("//third_party/github.com/bazelbuild/bazel-skylib:rules/copy_file.bzl", "co
 genrule(
     name = "generate_README",
     srcs = [
-        "@npm_bazel_terser//:index.md",
-        "//packages/terser/docs:install.md",
+        "index.md",
+        "install.md",
     ],
     outs = ["README.md"],
-    cmd = """cat $(execpath //packages/terser/docs:install.md) $(execpath @npm_bazel_terser//:index.md) | sed 's/^##/\\\n##/' > $@""",
+    cmd = """cat $(execpath install.md) $(execpath index.md) | sed 's/^##/\\\n##/' > $@""",
     tags = ["fix-windows"],
     visibility = ["//docs:__pkg__"],
 )
@@ -39,19 +63,17 @@ copy_file(
 pkg_npm(
     name = "npm_package",
     srcs = [
-        "@npm_bazel_terser//:package_contents",
+        "index.bzl",
+        "index.js",
+        "package.json",
+        "terser_config.default.json",
+        "terser_minified.bzl",
     ],
     build_file_content = """exports_files(["terser_config.default.json"])""",
-    substitutions = {
-        "@npm_bazel_terser//:terser-local": "@npm//@bazel/terser/bin:terser",
-    },
-    vendor_external = [
-        "npm_bazel_terser",
-    ],
     deps = [
         ":npm_version_check",
     ] + select({
-        # FIXME: fix stardoc on Windows; @npm_bazel_terser//:index.md generation fails with:
+        # FIXME: fix stardoc on Windows; //packages/terser:index.md generation fails with:
         #   ERROR: D:/b/62unjjin/external/npm_bazel_terser/BUILD.bazel:32:1: Couldn't build file
         #   external/npm_bazel_terser/docs.raw: Generating proto for Starlark doc for docs failed (Exit 1)
         "@bazel_tools//src/conditions:windows": [],
