+++ packages/terser/bin/BUILD.bazel
@@ -1,0 +1,14 @@
+"""The default binaries used in our local builds
+
+Note that in the npm distribution, this file is absent, and will be generated on-the-fly when
+the package is installed (possibly with different attribute values).
+"""
+
+load("@build_bazel_rules_nodejs//:index.bzl", "nodejs_binary")
+
+nodejs_binary(
+    name = "terser",
+    data = ["@npm//terser"],
+    entry_point = "//packages/terser:index.js",
+    visibility = ["//:__subpackages__"],
+)
