--- packages/terser/docs/BUILD.bazel
@@ -1,3 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-exports_files(["install.md"])
