+++ packages/terser/index.js
