+++ packages/terser/install.md
@@ -8,15 +8,9 @@ Wraps the Terser CLI documented at https://github.com/terser-js/terser#command-l
 
 Add the `@bazel/terser` npm package to your `devDependencies` in `package.json`.
 
-Your `WORKSPACE` should declare a `yarn_install` or `npm_install` rule named `npm`.
-It should then install the rules found in the npm packages using the `install_bazel_dependencies` function.
-See https://github.com/bazelbuild/rules_nodejs/#quickstart
-
-This causes the `@bazel/terser` package to be installed as a Bazel workspace named `npm_bazel_terser`.
-
 ## Installing with self-managed dependencies
 
-If you didn't use the `yarn_install` or `npm_install` rule to create an `npm` workspace, you'll have to declare a rule in your root `BUILD.bazel` file to execute terser:
+If you didn't use the `yarn_install` or `npm_install` rule, you'll have to declare a rule in your root `BUILD.bazel` file to execute terser:
 
 ```python
 # Create a terser rule to use in terser_minified#terser_bin
