+++ packages/terser/package.json
@@ -21,11 +21,6 @@
     "bin": {
         "terser": "index.js"
     },
-    "bazelWorkspaces": {
-        "npm_bazel_terser": {
-            "rootPath": "."
-        }
-    },
     "scripts": {
       "postinstall": "node npm_version_check.js"
     }
