--- packages/terser/src/BUILD.bazel
@@ -1,56 +1,0 @@
-# Copyright 2018 The Bazel Authors. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
-load("@build_bazel_rules_nodejs//:index.bzl", "nodejs_binary")
-load("@build_bazel_rules_nodejs//tools/stardoc:index.bzl", "stardoc")
-
-package(default_visibility = ["//visibility:public"])
-
-exports_files(["terser_config.default.json"])
-
-bzl_library(
-    name = "bzl",
-    srcs = glob(["*.bzl"]),
-    deps = [
-        "@build_bazel_rules_nodejs//:bzl",
-        "@build_bazel_rules_nodejs//internal/common:bzl",
-    ],
-)
-
-stardoc(
-    name = "docs",
-    testonly = True,
-    out = "index.md",
-    input = "index.bzl",
-    tags = ["fix-windows"],
-    deps = [":bzl"],
-)
-
-filegroup(
-    name = "package_contents",
-    srcs = [
-        "index.bzl",
-        "index.js",
-        "package.json",
-        "terser_config.default.json",
-        "terser_minified.bzl",
-    ],
-)
-
-nodejs_binary(
-    name = "terser-local",
-    data = ["@npm//terser"],
-    entry_point = "index.js",
-)
