--- packages/terser/src/WORKSPACE
@@ -1,1 +1,0 @@
-workspace(name = "npm_bazel_terser")
