+++ packages/terser/terser_config.default.json
