+++ packages/terser/test/directory_input/BUILD.bazel
@@ -1,6 +1,6 @@
 load("@build_bazel_rules_nodejs//:tools/declare_directory.bzl", "declare_directory")
 load("@npm_bazel_jasmine//:index.bzl", "jasmine_node_test")
-load("@npm_bazel_terser//:index.bzl", "terser_minified")
+load("//packages/terser:index.bzl", "terser_minified")
 
 declare_directory(
     name = "dir",
