+++ packages/terser/test/inline_sourcemap/BUILD.bazel
@@ -1,5 +1,5 @@
 load("@npm_bazel_jasmine//:index.bzl", "jasmine_node_test")
-load("@npm_bazel_terser//:index.bzl", "terser_minified")
+load("//packages/terser:index.bzl", "terser_minified")
 
 # Check that filegroups work
 filegroup(
